% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\name{collect}
\alias{collect}
\title{Force computation of a database query}
\arguments{
\item{x}{Lazy data frame backed by a database query.}

\item{...}{Additional arguments passed to methods.}
}
\description{
See \code{\link[dplyr:compute]{dplyr::collect()}} for details.
}
