% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_add_date}
\alias{db_add_date}
\title{Create new field with local date}
\usage{
db_add_date(tbl, time_col = "Observed_Date_Time", date_col = NULL)
}
\arguments{
\item{tbl}{DuckDB tbl created with \code{\link[=renmods_connect]{renmods_connect()}}}

\item{date_col}{Character. Name of the new field/column to create.}

\item{time_name}{Character. Field/Column name for date/time column from which
to extract a date.}
}
\value{
A \code{tbl_duckdb_connection} object - a lazy DuckDB table. With the
additional field.
}
\description{
Converts a time/date column to just local date (i.e. truncates the time
and converts to date).
}
