% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_fmt_times}
\alias{db_fmt_times}
\title{Convert to date/time}
\usage{
db_fmt_times(tbl)
}
\arguments{
\item{tbl}{DuckDB tbl created with \code{\link[=renmods_connect]{renmods_connect()}}}
}
\value{
A \code{tbl_duckdb_connection} object - a lazy DuckDB table. With the
converted fields.
}
\description{
Identifies and converts character date/time fields to date time.
}
