% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{renmods_disconnect}
\alias{renmods_disconnect}
\title{Disconnect from the database}
\usage{
renmods_disconnect(tbl)
}
\arguments{
\item{tbl}{DuckDB tbl created with \code{\link[=renmods_connect]{renmods_connect()}}}
}
\value{
\code{NULL} invisibly. Called for side effect of disconnecting database.
}
\description{
Disconnect from the database
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
db <- renmods_connect()
renmods_disconnect(db)
\dontshow{\}) # examplesIf}
}
