% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renmods-package.R
\docType{package}
\name{renmods-package}
\alias{renmods}
\alias{renmods-package}
\title{renmods: Downloads and queries data from B.C. Environmental Monitoring Data System (EnMoDs)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/renmods}
  \item \url{https://bcgov.github.io/renmods/}
  \item Report bugs at \url{https://github.com/bcgov/renmods/issues}
}

}
\author{
\strong{Maintainer}: Steffi LaZerte \email{sel@steffilazerte.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
